requirements.txt
git+https://github.com/MTG/freesound-python.git
ffmpeg-python==0.2.0